Fixing typo, autolayout warnings, moving styles to base styles
